Drop unnecessary touch commands

+++ compdevkit/cli.py
@@ -35,15 +35,14 @@ installtemplate = "# bash commands for installing your package"
 
 
 def write_template(path, template):
-    with open(path, "w") as f:
-        f.write(template)
+    if not path.exists():
+        with open(path, "w") as f:
+            f.write(template)
 
 def init():
     comptoplevel = Path.cwd() / "compconfig"
     comptoplevel.mkdir(exist_ok=True)
-    (comptoplevel / "setup.py").touch()
     write_template(comptoplevel / "setup.py", setuptemplate)
-    (comptoplevel / "install.sh").touch()
     write_template(comptoplevel / "install.sh", installtemplate)
 
     comp = comptoplevel / "compconfig"
@@ -51,7 +50,6 @@ def init():
 
     (comp / "__init__.py").touch()
 
-    (comp / "functions.py").touch()
     write_template(comp / "functions.py", functionstemplate)
 
     test = comp / "tests"
@@ -59,5 +57,4 @@ def init():
 
     (test / "__init__.py").touch()
 
-    (test / "test_functions.py").touch()
     write_template(test / "test_functions.py", testfunctionstemplate)
